+++ HW06/src/main/java/hw06/task1/services/ProductService.java
@@ -15,6 +15,6 @@ public interface ProductService {
     List<Product> findAll();
     Product findById(Integer id);
     List<Product> findByMaxUseBefore(int useBefore);
-    List<Product> findByMinPrice(Float price);
+    List<Product> findByMinPrice(int price);
     List<Product> findByBestBeforeDate(int manufactured, int useBefore);
 }
